Moved PALM code from prompt.go to palm.go
